+++ src/sec_adapter_signature.c
@@ -145,10 +145,10 @@ Sec_Result SecSignature_Process(Sec_SignatureHandle* signatureHandle, SEC_BYTE* 
         const Sec_Key* key = get_key(signatureHandle->keyHandle);
         // Get the out_length since it is not given to us.
         sa_status status = sa_invoke(signatureHandle->processorHandle, SA_CRYPTO_SIGN, NULL, &out_length,
-                signature_algorithm, key->handle, input, inputSize, parameters);
+                signature_algorithm, key->handle, input, (size_t)inputSize, parameters);
         CHECK_STATUS(status)
         status = sa_invoke(signatureHandle->processorHandle, SA_CRYPTO_SIGN, signature, &out_length,
-                signature_algorithm, key->handle, input, inputSize, parameters);
+                signature_algorithm, key->handle, input, (size_t)inputSize, parameters);
         CHECK_STATUS(status)
         *signatureSize = out_length;
     } else {
